requirements.txt
absl-py==1.2.0
apparmor==3.1.1
appdirs==1.4.4
application-utility==1.3.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
astunparse==1.6.3
async-generator==1.10
attrs==21.4.0
autocommand==2.2.1
backcall==0.2.0
bcrypt==4.0.0
beautifulsoup4==4.11.1
bidict==0.22.0
bleach==5.0.1
btrfsutil==5.19
CacheControl==0.12.11
cachetools==5.2.0
ceph==1.0.0
ceph-volume==1.0.0
cephfs==2.0.0
cephfs-shell==0.0.1
certifi==2022.6.15
cffi==1.15.1
chardet==5.0.0
click==8.1.3
cmd2==2.4.2
colorama==0.4.5
commonmark==0.9.1
contextlib2==21.6.0
cryptography==37.0.4
cupshelpers==1.0
Cython==0.29.32
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
distro==1.7.0
Django==4.0.6
dnspython==2.2.1
docopt==0.6.2
docutils==0.19
entrypoints==0.4
espeakng==1.0.2
executing==0.8.3
fastjsonschema==2.15.3
Flask==2.0.2
flatbuffers==2.0.7
future==0.18.2
gast==0.4.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.1
h11==0.13.0
h5py==3.7.0
html5lib==1.1
idna==3.3
importlib-metadata==4.8.3
imutils==0.5.4
inflect==6.0.0
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
itsdangerous==2.1.2
jaraco.context==4.1.2
jaraco.functools==3.5.1
jaraco.text==3.9.1
jedi==0.18.1
Jinja2==3.1.2
jira==3.3.1
jsonschema==4.6.0
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
keras==2.10.0
Keras-Preprocessing==1.1.2
keyutils==0.6
LibAppArmor==3.1.1
libclang==14.0.6
libtorrent==2.0.7
lit==14.0.6.dev0
louis==3.23.0
manjaro-sdk==0.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
more-itertools==8.13.0
msgpack==1.0.4
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
netsnmp-python==1.0a1
notebook==6.4.12
npyscreen==4.10.5
nspektr==0.4.0
numpy==1.22.4
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
ordered-set==4.1.0
outcome==1.2.0
packaging==21.3
pacman-mirrors==4.23.2
pandocfilters==1.5.0
parso==0.8.3
pep517==0.13.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
ply==3.11
progress==1.6
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.12
pycairo==1.21.0
pycparser==2.21
pycups==2.0.1
pycurl==7.45.1
pydantic==1.10.2
pygame==2.1.2
Pygments==2.12.0
PyGObject==3.42.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyperclip==1.8.2
PyQt5==5.15.7
PyQt5-sip==12.11.0
pyrsistent==0.18.1
PySide6==6.3.1
PySimpleGUI==4.60.3
pysmbc==1.0.23
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
python-engineio==4.3.3
python-socketio==5.7.0
pyttsx3==2.90
PyYAML==6.0
pyzbar==0.1.9
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
rados==2.0.0
rbd==2.0.0
reportlab==3.6.11
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
resolvelib==0.8.1
retrying==1.3.3
rgw==2.0.0
rsa==4.9
selenium==4.3.0
Send2Trash==1.8.0
shiboken6==6.3.1
six==1.16.0
sniffio==1.2.0
socket-client-server==0.5.dev3
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
srt==3.5.2
stack-data==0.3.0
systemd-python==235
team==1.0
tenacity==8.0.1
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.15.0
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tornado==6.1
tqdm==4.64.1
traitlets==5.3.0
trio==0.21.0
trio-websocket==0.9.2
trove-classifiers==2022.8.31
typing_extensions==4.3.0
udiskie==2.4.2
urllib3==1.26.12
validate-pyproject==0.10.1
vosk==0.3.43
Wave==0.0.2
wcwidth==0.2.5
webdriver-manager==3.8.2
webencodings==0.5.1
Werkzeug==2.1.2
widgetsnbextension==3.6.1
wrapt==1.14.1
wsproto==1.1.0
zipp==3.8.1
